Changed hover state of buttons and position of start and next button
The hover state has been changed by removing the scale increase and instead having the background colour increase brighteness. The next button would appear and reposition all the answer buttons. Now, with position absolute and margin top, the start and next button appears underneath the container and has a much better user experience as it doeasn't effect the position of the other elements once it appears.